backend/requirements.txt
# Azure SDKs
azure-core>=1.30.1
azure-ai-documentintelligence==1.0.0b2
openai>=1.14.3

# Logging & env
python-dotenv>=1.0.0
tqdm>=4.66.0
python-slugify>=8.0.0

# PDF handling
pypdf>=3.10.0

# Web framework (backend)
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
python-multipart>=0.0.9
